Move node files for UI testing to v1.0 directory structures
