+++ test/editor/pageobjects/nodes/core/common/21-debug_page.js
