+++ test/editor/pageobjects/nodes/core/function/10-function_page.js
