+++ test/editor/pageobjects/nodes/core/function/15-change_page.js
