+++ test/editor/pageobjects/nodes/core/function/16-range_page.js
