+++ test/editor/pageobjects/nodes/core/function/80-template_page.js
